Use extraSupportRotation in wooden DrawSupportForSequence functions
This uses the `extraSupportRotation` in the track element data in the wooden `DrawSupportForSequenceA/B` functions.

It removes the down transition types as they are no longer necessary as they are rotated by this instead. They were only used in the track element data, which is only used by these functions.

It fills in some of the track element data necessary to draw all the ones currently used by these functions.

This should do nothing as is. I have diffed pictures of all track pieces of all affected track types and found no difference.

This is so that you will be able to use rotated transition types, other than the two up and down rotations previously available, on track pieces that require them like sloped turns. It's not strictly necessary but it would be nice to be able to use this function for supports on the quarter helices.

This also adds a const for the count of transition types, and renames a data array to the current style.