Update process execution flow for Linux (#25206)
Previously zenity/kdialog were launched in a way that would prevent
events from getting pumped to the new application in some scenarios.

This new approach, mimicked from
https://github.com/libsdl-org/SDL/blob/9859c051781771015356e31f941bab40a763bea9/src/video/wayland/SDL_waylandmessagebox.c#L40-L75
fixes this problem for compositors where it used to be a problem.

In particular, this fixes the system dialog not launching on the Steam
Deck.